fix(deployment): fix https://github.com/CleverseAcademy/cd-compose-deployment/issues/8

+++ usecases/interface.go
@@ -22,7 +22,7 @@ type EventLogUseCase struct {
 
 type (
 	IUseCasePrepareServiceDeployment interface {
-		Execute(serviceName entities.ServiceName, priotity int8, ref string, image string) (*entities.Deployment, error)
+		Execute(serviceName entities.ServiceName, priotity uint64, ref string, image string) (*entities.Deployment, error)
 	}
 	IUseCaseEnqueueServiceDeployment interface {
 		Execute(entities.ServiceName, *entities.Deployment) int8
